% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdi_group_by.R
\name{sbdi_group_by}
\alias{sbdi_group_by}
\title{Specify fields to group when downloading record counts}
\usage{
sbdi_group_by(...)
}
\arguments{
\item{...}{zero or more individual column names to include}
}
\value{
If any arguments are provided, returns a `data.frame` with
columns `name` and `type`, as per [select.data_request()].
}
\description{
`count.data_request()` and `atlas_counts()` support server-side grouping of
data. Grouping can be used to return record counts grouped by multiple, valid
fields (found by `search_all(fields)`).
}
\examples{
\dontrun{
sbdi_call() |>
  sbdi_group_by(basisOfRecord) |>
  atlas_counts()
}
}
