% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdi_identify.R
\name{sbdi_identify}
\alias{sbdi_identify}
\title{Narrow a query by passing taxonomic identifiers}
\usage{
sbdi_identify(...)
}
\arguments{
\item{...}{One or more scientific names.}
}
\value{
A tibble containing identified taxa.
}
\description{
When conducting a search or creating a data query, it is common to identify
a known taxon or group of taxa to narrow down the records or results returned.
}
\details{
`sbdi_identify()` is used to identify taxa you want returned in a search or
a data query. Users to pass scientific names or taxonomic identifiers
with pipes to provide data only for the biological group of interest.

It is good to use [search_taxa()] and [search_identifiers()]
first to check that the taxa you provide to `sbdi_identify()` return the
correct results.
}
\examples{
\dontrun{
# Specify a taxon. A valid taxon will return an identifier.
sbdi_identify("reptilia")

# Specify more than one taxon at a time.
sbdi_identify("reptilia", "mammalia", "aves", "pisces")

# Use `sbdi_identify()` to narrow your queries
sbdi_call() |>
  sbdi_identify("Eolophus") |>
  atlas_counts()

# Within a pipe, `identify()` and `sbdi_identify()` are synonymous.
# hence the following is identical to the previous example:
request_data() |>
  identify("Eolophus") |>
  count() |>
  collect()

# If you know a valid taxon identifier, use `sbdi_filter()` instead.
id <- "https://biodiversity.org.au/afd/taxa/009169a9-a916-40ee-866c-669ae0a21c5c"
sbdi_call() |>
  sbdi_filter(lsid == id) |>
  atlas_counts()
}
}
\seealso{
[search_taxa()] to find identifiers from scientific names;
[search_identifiers()] for how to get names if taxonomic identifiers
are already known.
}
