% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe.R
\docType{data}
\name{swe}
\alias{swe}
\alias{swe_wgs84}
\title{Basic Sweden polygons for visualization}
\format{
list of Simple Feature Collection (package sf) CRS: EPSG:3006
}
\source{
Statistikmyndigheten SCB \url{https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser}
}
\usage{
data(swe)

data(swe_wgs84)
}
\description{
A basic polygon data set in Simple Feature Collection (package sf) representing
Sweden.

A basic polygon data set in Simple Feature Collection (package sf) representing
Sweden.
}
\keyword{data}
\keyword{sets}
