% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden_grid.R
\docType{data}
\name{Sweden_Grid}
\alias{Sweden_Grid}
\alias{Sweden_Grid_5km_Wgs84}
\alias{Sweden_Grid_10km_Wgs84}
\alias{Sweden_Grid_25km_Wgs84}
\alias{Sweden_Grid_50km_Wgs84}
\title{The standard Swedish squared grid}
\format{
Simple Feature Collection (package sf) CRS: EPSG:4326
}
\source{
https://github.com/biodiversitydata-se/Sweden_Grids
}
\usage{
data(Sweden_Grid_5km_Wgs84)

data(Sweden_Grid_10km_Wgs84)

data(Sweden_Grid_25km_Wgs84)

data(Sweden_Grid_50km_Wgs84)
}
\description{
A 5, 10, 25 and 50 km wide grid data set in Simple Feature Collection (package sf) over Sweden.
}
\keyword{datasets}
